chatbot/app/requirements.txt
fastapi
uvicorn
pillow
torch
timm
pymilvus
python-multipart
openai